pyproject.toml
[project]
name = "Template_Django"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "", email = ""},
]
dependencies = [
    "django>=5.0.7",
    "djangorestframework>=3.15.2",
    "python-dotenv>=1.0.1",
    "passage-identity>=2.1.4",
]
requires-python = "==3.10.*"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm]
distribution = false

[tool.pdm.scripts]
dev = "python manage.py runserver localhost:8000"
runserver = "python manage.py runserver"
createsuperuser = "python manage.py createsuperuser"
migrate = "python manage.py migrate"
makemigrations = "python manage.py makemigrations"
model = "pdm run python manage.py graph_models -a -o modelagem.png"